web-backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
Django==4.0.5
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
idna==3.3
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
random-username==1.0.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
URLSearchParams==1.2
django-storages==1.12.3
boto3==1.24.29
drf_yasg==1.21.3